Refactor deprecated methods used for annotators in favor of new api Remove old features to display line markers for coercions (on property and value arguments) and givens (injections) Some more refactor
