+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/ProofsIdePlugin.kt
@@ -6,10 +6,8 @@ import arrow.meta.ide.invoke
 import arrow.meta.ide.plugins.proofs.annotators.coercionAnnotator
 import arrow.meta.ide.plugins.proofs.annotators.givenAnnotator
 import arrow.meta.ide.plugins.proofs.annotators.refinementAnnotator
-import arrow.meta.ide.plugins.proofs.coercions.coercionInspections
+import arrow.meta.ide.plugins.proofs.inspections.coercions.coercionInspections
 import arrow.meta.ide.plugins.proofs.folding.codeFolding
-import arrow.meta.ide.plugins.proofs.markers.givenCallLineMarker
-import arrow.meta.ide.plugins.proofs.markers.givenParamLineMarker
 import arrow.meta.ide.plugins.proofs.markers.proofLineMarkers
 import arrow.meta.ide.plugins.proofs.markers.refinementLineMarkers
 import arrow.meta.ide.plugins.proofs.psi.isCoercionProof
@@ -21,7 +19,6 @@ import arrow.meta.plugins.proofs.phases.resolve.diagnostics.suppressProvenTypeMi
 import org.jetbrains.kotlin.psi.KtClassOrObject
 import org.jetbrains.kotlin.psi.KtFunction
 import org.jetbrains.kotlin.psi.KtNamedFunction
-import org.jetbrains.kotlin.psi.KtParameter
 import org.jetbrains.kotlin.psi.KtProperty
 
 val IdeMetaPlugin.typeProofsIde: IdePlugin
