+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/inspections/coercions/coercionInspection.kt
@@ -1,8 +1,8 @@
-package arrow.meta.ide.plugins.proofs.coercions
+package arrow.meta.ide.plugins.proofs.inspections.coercions
 
 import arrow.meta.ide.IdeMetaPlugin
-import arrow.meta.ide.plugins.proofs.coercions.explicit.localExplicitCoercions
-import arrow.meta.ide.plugins.proofs.coercions.implicit.localImplicitCoercion
+import arrow.meta.ide.plugins.proofs.inspections.coercions.explicit.localExplicitCoercions
+import arrow.meta.ide.plugins.proofs.inspections.coercions.implicit.localImplicitCoercion
 import arrow.meta.phases.Composite
 import arrow.meta.phases.ExtensionPhase
