+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/predef.kt
@@ -1,24 +1,30 @@
-package arrow.meta.ide.plugins.proofs.coercions
+package arrow.meta.ide.plugins.proofs
 
 import arrow.meta.ide.dsl.utils.replaceK
 import arrow.meta.ide.plugins.proofs.markers.description
-import arrow.meta.ide.plugins.proofs.markers.markerMessage
-import arrow.meta.ide.plugins.proofs.markers.participatingTypes
 import arrow.meta.phases.CompilerContext
 import arrow.meta.plugins.proofs.phases.CoercionProof
+import arrow.meta.plugins.proofs.phases.areTypesCoerced
 import arrow.meta.plugins.proofs.phases.coerceProof
 import arrow.meta.quotes.ktFile
 import com.intellij.psi.PsiElement
+import com.intellij.psi.util.PsiTreeUtil
+import org.jetbrains.kotlin.descriptors.CallableDescriptor
 import org.jetbrains.kotlin.idea.caches.resolve.analyze
 import org.jetbrains.kotlin.idea.imports.importableFqName
 import org.jetbrains.kotlin.name.FqName
+import org.jetbrains.kotlin.nj2k.postProcessing.type
+import org.jetbrains.kotlin.psi.KtCallExpression
 import org.jetbrains.kotlin.psi.KtDotQualifiedExpression
+import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.kotlin.psi.KtImportDirective
 import org.jetbrains.kotlin.psi.KtImportList
 import org.jetbrains.kotlin.psi.KtProperty
 import org.jetbrains.kotlin.psi.KtValueArgument
 import org.jetbrains.kotlin.resolve.ImportPath
+import org.jetbrains.kotlin.resolve.calls.callUtil.getResolvedCall
+import org.jetbrains.kotlin.resolve.calls.model.ResolvedCall
 import org.jetbrains.kotlin.resolve.lazy.BodyResolveMode
 import org.jetbrains.kotlin.types.KotlinType
 
@@ -75,3 +81,33 @@ private fun CompilerContext.replaceWithProof(ktExpression: KtExpression, pairTyp
       ktExpression.replaceK(new)
     }
   }
+
+internal fun KtElement.participatingTypes(): Pair<KotlinType, KotlinType>? =
+  when (this) {
+    is KtProperty -> participatingTypes()
+    is KtValueArgument -> participatingTypes()
+    else -> null
+  }
+
+internal fun CompilerContext?.isCoerced(ktElement: KtElement): Boolean =
+  ktElement.participatingTypes()?.let { (subtype, supertype) ->
+    this.areTypesCoerced(subtype, supertype)
+  } ?: false
+
+private fun KtProperty.participatingTypes(): Pair<KotlinType, KotlinType>? {
+  val subType: KotlinType? = initializer?.resolveKotlinType()
+  val superType: KotlinType? = type()
+  return subType.pairOrNull(superType)
+}
+
+// TODO: can't resolve Pair<Type, Type> or other HKT * -> *
+private fun KtValueArgument.participatingTypes(): Pair<KotlinType, KotlinType>? {
+  val subType: KotlinType? = getArgumentExpression()?.resolveKotlinType()
+
+  val ktCallExpression: KtCallExpression? = PsiTreeUtil.getParentOfType(this, KtCallExpression::class.java)
+  val myselfIndex: Int = ktCallExpression?.valueArguments?.indexOf(this) ?: 0
+  val superType: KotlinType? = ktCallExpression.getResolvedCall(analyze())?.let { resolvedCall: ResolvedCall<out CallableDescriptor> ->
+    resolvedCall.resultingDescriptor.valueParameters.getOrNull(myselfIndex)?.type
+  }
+  return subType.pairOrNull(superType)
+}
